FUSETOOLS2-1316 - use LSP4E JDT Integration
Now that there is one, it avoids having the same done in our codebase.
All the more that it is duplicating completion proposal when both are
installed.

It contains also upgrades to Eclipse 2021-09 as there was strange errors
without it but I do not think it is worth y to investigate more.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>